+++ Angular-Front-End/app-front-end/src/app/components/user-profile-page/user-profile-page.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthenticationService } from 'src/app/services/authentication/authentication.service';
 import { UserModel } from 'src/app/models/user.model';
+import { UserCommunityService } from 'src/app/services/controllers/user-community-controller.service';
 
 @Component({
   selector: 'app-user-profile-page',
@@ -10,15 +11,35 @@ import { UserModel } from 'src/app/models/user.model';
 export class UserProfilePageComponent implements OnInit {
 
   currentUser: UserModel;
+  communities: any[];
+  checkCommunities: boolean = true;
+  error: any;
 
   constructor(
-    private authenticationService: AuthenticationService
+    private authenticationService: AuthenticationService,
+    private userCommunityService: UserCommunityService,
   ) { 
-    this.authenticationService.currentUser.subscribe(x => this.currentUser = x);
+    this.authenticationService.currentUser.subscribe(user => this.currentUser = user);
   }
 
   ngOnInit(): void {
+    this.userCommunityService.userSubbedCommunities(this.currentUser.id).subscribe((communityList: any) => {
+      this.communities = communityList.second.list;
+      if(!communityList.second.list || communityList?.second.list.length === 0) {
+        this.checkCommunities = false;
+      }
+    });
   }
 
+  unsubscribeFromCommunity(id: string): void {
+    this.userCommunityService.unsubscribeFromCommunity(this.currentUser.id, id).subscribe();
+
+    this.communities.splice(this.communities.indexOf(id), 1);
+
+    // check if there are still communities
+    if(!this.communities || this.communities?.length === 0) {
+      this.checkCommunities = false;
+    }
+  }
 
 }