+++ Angular-Front-End/app-front-end/src/app/models/community.model.ts
@@ -3,5 +3,5 @@ import { Identifiable } from './identifiable';
 export interface CommunityModel extends Identifiable{
     name: string;
     description: string;
-    menuShow?: boolean;
+    isSubscribed?: boolean;
 }