+++ Angular-Front-End/app-front-end/src/app/services/controllers/user-community-controller.service.ts
@@ -10,14 +10,24 @@ export class UserCommunityService {
     private httpOptions = {
         headers: new HttpHeaders({ "Content-Type": "application/json" })
     }
-    private subscribeToCommunityUrl = '/userCommunityApi/subscribeToCommunity'
+    private subscribeToCommunityUrl = '/userCommunityApi/subscribeToCommunity';
+    private unsubscribeFromCommunityUrl = '/userCommunityApi/unsubscribeFromCommunity';
+    private userSubbedCommunitiesUrl = '/userCommunityApi/userSubbedCommunities';
 
     constructor(
         protected http: HttpClient
       ) { }
 
-    public subscribeToCommunity(uID: string, cID: string) {
+    public subscribeToCommunity(uID: string, cID: string): any {
         return this.http.post(this.subscribeToCommunityUrl, {uID, cID});
+    }
+
+    public unsubscribeFromCommunity(uID: string, cID: string): any {
+        return this.http.post(this.unsubscribeFromCommunityUrl, {uID, cID});
+    }
+
+    public userSubbedCommunities(uID: string): any {
+        return this.http.get(`${this.userSubbedCommunitiesUrl}/${uID}`);
     }
 }