+++ changelog.md
@@ -3,6 +3,18 @@ All changes made to this project will be documented in this file
 
 ## [Unreleased]
 
+## [0.1.3] - 2020-08-06
+### Added
+- Subscribe and Unsubscribe features now fully work
+- 
+
+### Changed
+- Interface is now fixed
+- Finished page for user-profile
+- Home page after login changed link to User-Page
+
+-------------------------------------------------------------------------
+
 ## [0.1.3] - 2020-08-05
 ### Added
 - Role backend was completed